refactor(import): extract staged/bank hooks + StagedPane/BankPane (FINLYNQ-118 phase 4)
Behaviour-preserving extraction of the import/pending god-component.
page.tsx 1796 -> 1180 lines (~34%). NO data-fetching library
(FINLYNQ-115); the bespoke fetch/useState/useEffect/finally pattern is
kept verbatim.

- _types.ts: StagedRow/ParsedAnchorRow/StagedDetail + daysUntil/shiftDays.
- _hooks/use-staged.ts: useStagedImports() + useStagedDetail() + useBankLedger().
  The hooks own load-on-mount/-change + the data state; the page keeps the
  tightly-coupled orchestration (openId/selected/match-action callbacks/
  highlight neighborhoods) verbatim, threading the state + setters through.
- _components/: staged-list-view, reconcile-header (+ re-apply modal),
  bank-pane (DbPane rowActions), staged-pane (FilePane rowActions +
  suggestions header), unbound-picker-pane.

safeName null-guards retained at the moved accountOptions callsite; approve /
apply-rules / per-row delete / link / flag behaviour unchanged.
